task_handler/requirements.txt
requests
aws-psycopg2
python-dotenv